# Speed up the airfoil drag calculations