--- a/src/aero/cdsum.jl
+++ b/src/aero/cdsum.jl
@@ -93,8 +93,6 @@ function cdsum!(ac, imission, ip, icdfun)
 #- - - - - - - - - - - - - - - - - - - - - - - - - - - -
       if (icdfun==1) 
 #----- integrated across span for CDwing
-
-      # if(Ldebug) write(*,*) 'calling SURFCD2...'
       clpo,clps,clpt,
 	cdfw,cdpw,CDwing,CDover = surfcd2(wing,gammat,gammas,
                                     Mach,CL,CLhtail,Reco,
@@ -200,8 +198,6 @@ function cdsum!(ac, imission, ip, icdfun)
 
 #- - - - - - - - - - - - - - - - - - - - - - - - - - - -
 #---- induced CD
-#      if(Ldebug) write(*,*) '...calling CDITRP...'
-
       cditrp(para, wing, htail)
       CDi = para[iaCDi]
 
@@ -222,13 +218,7 @@ function cdsum!(ac, imission, ip, icdfun)
 #---- total CD
       CD = CDi + CDfuse + CDwing + CDover + CDhtail + CDvtail + CDstrut + CDnace + dCDBLIf + dCDBLIw
       para[iaCD] = CD      
-      CD_components = [CDi  CDfuse  CDwing  CDover CDhtail  CDvtail  CDstrut 	CDnace dCDBLIf dCDBLIw]
-      # println(CD_components)
 
-#- - - - - - - - - - - - - - - - - - - - - - - - - - - -
-#      if(Ldebug) write(*,*) '...exiting CDSUM...'
-      # println("Total CD = ", CD)
-      # println("PARA = ", para)
       return
 end # cdsum