… docs/src/aero/drag.md
… docs/src/aero/theory_airfun_and_splines.md
… test/aero.txt
… test/benchmark_elements.jl
… test/default_sized.jl
… test/default_structures.jl
… test/geom.txt
… test/regression_test_wsize.jl
… test/unit_test_aero.jl
… test/unit_test_loads.jl
… test/unit_test_structures.jl
… test/weights.txt