This PR makes some improvements to the airfoil drag calculations to speed up the airsection calculations by roughly ~5x. The airfoil type wasn't really type stable/ it's field types weren't inferable by the compiler. Also the spline evaluations have been cleaned up a little and optimized for better performance.

On my M2 mac the performance of the airfoil section calculations are:
```julia
julia> bench = @benchmarkable aerodynamics.airfun($clp, $toc, $Mperp, 
                                               $airfoil_section) seconds = 30 evals = 100
Benchmark(evals=100, seconds=30.0, samples=10000)

julia> bench_airfun = run(bench)
BenchmarkTools.Trial: 10000 samples with 100 evaluations.
 Range (min … max):  696.250 ns …  1.106 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     701.670 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   710.746 ns ± 24.497 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▇█▇▆▅▅▄▂▂▂▁▁    ▃▄▃▃▂▁                                       ▂
  █████████████▇▇██████████▆▇▇▆▆▅▆▆▆▇▅▇▆▆▆▅▆▄▆▅▆▅▄▅▄▃▅▄▄▄▅▃▄▂▅ █
  696 ns        Histogram: log(frequency) by time       824 ns <

 Memory estimate: 96 bytes, allocs estimate: 4.
```

This PR:
```julia
julia> bench_airfun = run(bench)
BenchmarkTools.Trial: 10000 samples with 100 evaluations.
 Range (min … max):  131.250 ns … 392.500 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     133.750 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   134.427 ns ±   7.138 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

              ▂  ▆  █  ▆  ▄                                      
  ▂▁▁▃▁▁▃▁▁▆▁▁█▁▁█▁▁█▁▁█▁▁█▁▁█▁▁▅▁▁▄▁▁▃▁▁▂▁▁▂▁▁▂▁▁▂▁▁▂▁▁▂▁▁▂▁▁▂ ▃
  131 ns           Histogram: frequency by time          140 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```