@askprash 2025-02-22
Got rid of a few more stray allocations that weren't needed. Wing cd calculations are allocation free now and significantly faster relative to [main](https://github.com/MIT-LAE/TASOPT.jl/commit/e182855856fc03eaa8950e8b317519bf66f82054):
```julia
julia> bench_surfcd2 = run(bench)
BenchmarkTools.Trial: 10000 samples with 5 evaluations.
 Range (min … max):  8.225 μs …  1.750 ms  ┊ GC (min … max): 0.00% … 99.12%
 Time  (median):     8.542 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   9.639 μs ± 29.016 μs  ┊ GC (mean ± σ):  7.69% ±  2.78%

  ▄▇█▇▅▅▄▃▄▄▃▃▂▂▂▂▁▁▁▁▁▁                                     ▂
  ██████████████████████████▇▇▇▆▅▅▆▅▄▆▆▄▅▅▄▄▅▁▅▄▅▁▄▄▄▅▃▄▄▄▃▄ █
  8.22 μs      Histogram: log(frequency) by time     14.6 μs <

 Memory estimate: 8.31 KiB, allocs estimate: 281.
```

This PR:
```julia
---------------------------------------
surfcd2 (FORTRAN on MacPro M2 ~ 2.01 μs)
---------------------------------------
BenchmarkTools.Trial: 10000 samples with 5 evaluations.
 Range (min … max):  1.633 μs …  12.058 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.750 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.782 μs ± 342.426 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▂▂▂▄█                                                       
  ▃█████▇▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂ ▃
  1.63 μs         Histogram: frequency by time        3.18 μs <

 Memory estimate: 0 bytes, allocs estimate: 0. 
```

New documentation added too.