@askprash 2025-02-22
Found a bug 🪲. The spline function had an error here: https://github.com/MIT-LAE/TASOPT.jl/blob/277bc6d29631b12374503643bc21f3357f720145/src/utils/spline.jl#L36